Add detail page
Key Changes:
- Use Hero for animating transition to detail page
- Load a larger version of the backdrop as the image in detail page
- Display more movie details such as ratings and the full overview
- Add tests

+++ lib/domain/entities/search_movie_info.dart
@@ -42,8 +42,8 @@ class SearchMovieInfo with EquatableMixin {
     releaseDate = json['release_date'];
     video = json['video'];
     adult = json['adult'];
-    popularity = json['popularity'];
-    voteAverage = json['vote_average'];
+    popularity = json['popularity'].toDouble();
+    voteAverage = json['vote_average'].toDouble();
     voteCount = json['vote_count'];
     id = json['id'];
